Fix: 로드 버튼 버그 수정
로드 버튼 여러 번 누를 시 리퀘스트가 셀 수만큼 발생하는 현상 수정
